wwt_data_formats/tests: gather up some test infrastructure
